refactor schema check
